[Task - 39] New home gesture to change between grid2, 3 and 4

+++ data/src/main/kotlin/jsanzo/movies/data/model/DataMovieDetails.kt
@@ -20,7 +20,7 @@ data class DataMovieDetails(
     val productionCompanies: List<DataMovieProductionCompany>,
     val productionCountries: List<DataMovieProductionCountry>,
     val releaseDate: String,
-    val revenue: Int,
+    val revenue: Long,
     val runtime: Int?,
     val spokenLanguages: List<DataMovieSpokenLanguage>,
     val status: String,
